+++ src/Objects/Venue.php
@@ -10,6 +10,9 @@ namespace Telegram\Bot\Objects;
  * @method string      getTitle()           Name of the venue.
  * @method string      getAddress()         Address of the venue.
  * @method string      getFoursquareId()    (Optional). Foursquare identifier of the venue.
+ * @method string      getFoursquareType()  (Optional). Foursquare type of the venue. (For example,
+ *                                          “arts_entertainment/default”, “arts_entertainment/aquarium” or
+ *                                          “food/icecream”.)
  */
 class Venue extends BaseObject
 {